Refactor pupil status function in preparation for a new version 2 message
